Introduce interactivity to visual diff tests (experimental)
